rename alt to altInstance, fix paths
