+++ app/actions/TopicActions.js
@@ -1,4 +1,4 @@
-import alt from 'alt';
+import alt from 'altInstance';
 import TopicWebAPIUtils from 'utils/TopicWebAPIUtils';
 
 /*
