+++ app/app.js
@@ -1,5 +1,5 @@
-import isomorphicRouterRenderer from './utils/IsomorphicRouterRenderer';
-import alt from 'alt';
-import routes from 'routes';
+import isomorphicRouterRenderer from 'utils/IsomorphicRouterRenderer';
+import alt from 'altInstance';
+import routes from './routes.js';
 
 export default isomorphicRouterRenderer(alt, routes);
