+++ app/utils/IsomorphicRouterRenderer.js
@@ -30,9 +30,7 @@ import Iso from 'iso';
 import React from 'react';
 import Router from 'react-router';
 
-import routes from 'routes';
-
-export default function IsomorphicRouterRenderer(alt) {
+export default function IsomorphicRouterRenderer(alt, routes) {
   let renderedMarkup;
   if (typeof window === 'undefined') {
     renderedMarkup = (state, url) => {
